+++ NEWS.md
@@ -2,6 +2,8 @@
 
  - bugfixes
      - contrast attributes are no longer dropped by `extend`
+     - more bugfixes and unit tests for binomial responses
+     - warnings for non-uniform weights, which aren't supported yet
 
 ### simr 1.0.3